Django/Blog/pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["AyaMoeen <ayamoinn95@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "5.1"
mypy = "^1.13.0"
ruff = "^0.7.2"
pytest = "^8.3.3"
djangorestframework = "^3.15.2"
django-stubs = "^5.1.1"
better-profanity = "^0.7.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
slugify = "^0.0.1"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
redis = "^5.2.0"
django-celery-results = "^2.5.1"
djangorestframework-simplejwt = "^5.3.1"
types-requests = "^2.32.0.20241016"
djangorestframework-stubs = "^3.15.1"
setuptools = "^75.6.0"
django-notifications-hq = "^1.8.3"
openai = "^1.57.0"
django-environ = "^0.11.2"
pydantic = "^2.10.3"
transformers = "^4.47.0"
torch = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
